==> 01b_chemical_sort_sensitivity.Rmd <==
---
title: "01b_chemical_sort_sensitivity"
author: "Taiba"
date: "2023-06-29"
output: pdf_document
---
############################################################################################################################################################################################# 
#Don't use this analysis it was only done for after committee meeting
```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, plotly)
```
#06/29/2023
# Step 1 Original raw pesticide dataset for 48 US States from 01_chemicals_sortingRmd file
```{r}
load("~/Desktop/Jabbeen/jt_dissert/jt_dissert/pest_conus_v2.RDA")
dat_conus$year<- as.numeric(as.character(dat_conus$year))
#write_csv(dat, "dat_pest_conusv1.csv")
```

#07/03/23 time series data
```{r}
# no of observations in dataset
length(dat_conus)

#first and last 10 obs
head(dat_conus, n=10)
tail(dat_conus, n=10)

#Basic time series plot
plot(dat_conus, xlab= "year", ylab= "pest_applied", main= "Pesticide applied in lbs in CONUS 1992-2019")

plot(dat_conus)
start(dat_conus)
end(dat_conus)

#use time(), deltat(), frequency(), cycle() with conus data for sampling frequency

#building time series objects
time_series <- ts(dat, start = , frequency = )
  
#check if object is ts class
is.ts(dat_conus)
  
# removing trends via log() and plotting linear using ts.plot()
x <- log(a)

ts.plot(dat)

#remove levels by differencing diff()
# remove seasonal trends with seasonal diff 
diff(dat, lag= s)

#white-noise (WN) model with fixed mean/variance
#no clear pattern/correlation using arima(dat, order= c(0,0,0))

#random walk (RW) model
#stationary process
#convert WN to RW using cumsum()

#scatterplot


```


# Step 1.1 creating two datasets for 14 yrs period separately from original 28 yr data
```{r}
dat1 <- dat_conus |>
  filter(year >= 1992 & year <=2005) #years 1992 to 2005 has 4585122 obs
# save same file in rda format as backup
save(dat1, file = "dat1_1992_2005.RDA")

dat2 <- dat_conus |>
  filter(year >= 2006 & year <=2019) #years 2006 to 2019 has 4977092 obs
# save same file in rda format as backup
save(dat2, file = "dat2_2006_2019.RDA")

```

# Step 2a creating variables median, temporal and spatial variables using original above dat1 (1992-2005)
```{r}
#median quantity and temporal coverage of chemicals applied
median_pest <- dat1 %>%
  group_by(state_name, county_fips_code, compound) %>%
  summarize(median_pest = median(pest_applied, na.rm = TRUE),
            n_yrs_repeat = sum(!is.na(pest_applied)),
            years = list(unique(year[!is.na(pest_applied)])))|>
  mutate(pct_yrs= round((n_yrs_repeat/14)*100, 2))


#spatial coverage using states (percent states with any compound applied per year)
state_cover <- median_pest %>%
  group_by(compound) %>% #431 unique compounds
  summarize(state_cvr = n_distinct(state_name[!is.na(median_pest)])) |>
  mutate(pct_state_cvr = round((state_cvr/48)*100, 2))

#join meadian_pest and state-cover
dat_med_st_cvr<- left_join(median_pest, state_cover, by=c("compound"))

######################################################################
#Spatial coverage of chemicals applied (coverage that shows percent counties)
summ <- median_pest %>%
    group_by(state_name, compound) %>%
    summarize(n_counties_pest = ifelse(any(!is.na(median_pest)),
                                       n_distinct(county_fips_code[!is.na(median_pest)]), 0))

#number of counties
n_cnty<- median_pest %>%
  group_by(state_name) %>%
  summarize(num_counties = n_distinct(county_fips_code))

#join n_cnty with spatial coverage to get percent spatial coverage
summ_spat_pct<- left_join(summ,n_cnty, by="state_name") |>
  mutate(pct_spat_cover = round((n_counties_pest/num_counties)*100, 2))
  

#########################################################################

# Final data set that has median pesticide applied, pct spatial and temporal coverage for dat1 (1992-2005)

chem_appl_9205<- left_join(dat_med_st_cvr, summ_spat_pct, 
                      by=c("state_name","compound"))

chem_appl_9205<- chem_appl_9205 |>
  rename(state=state_name, pct_state_cvr = pct_state_cvr,
         pct_cnty_cover=pct_spat_cover) |>
  mutate(years = gsub("(^c\\(|\\)$)", "", years))

# save same file in rda format as backup
save(chem_appl_9205, file = "chem_appl_spat_temp_cvr92_2005.RDA")
```


#Step 3a chemical selection function
```{r}
compound_filter <- function(data, pct_yrs_thresh, pct_state_cvr_thresh, pct_cnty_cvr_thresh) {
  
  # Filter the data based on pct_yrs threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_yrs) >= pct_yrs_thresh)
  
  # Filter the data based on pct_state_cvr threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_state_cvr) >= pct_state_cvr_thresh)
  
  # Filter the data based on pct_cnty_cvr threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_cnty_cover) >= pct_cnty_cvr_thresh)
  
  # Return unique list of compounds that match the criteria as a dataframe
  unique(data$compound) %>% as.data.frame()
}
```

# Different percentages generate different lists
```{r}
#Use file: load("chem_appl_spat_temp_cvr92_2005.RDA")
compound_filter<- compound_filter(chem_appl_9205, 80, 80, 80)
compound_filter<- compound_filter |> rename(compound = ".")
#write_csv(compound_filter, "chemical_list90_90_90.csv")
```

########################################################################################################################################################################################

# Step 2b creating variables median, temporal and spatial variables using original above dat1 (2006-2019)
```{r}
#median quantity and temporal coverage of chemicals applied
median_pest <- dat2 %>%
  group_by(state_name, county_fips_code, compound) %>%
  summarize(median_pest = median(pest_applied, na.rm = TRUE),
            n_yrs_repeat = sum(!is.na(pest_applied)),
            years = list(unique(year[!is.na(pest_applied)])))|>
  mutate(pct_yrs= round((n_yrs_repeat/14)*100, 2))


#spatial coverage using states (percent states with any compound applied per year)
state_cover <- median_pest %>%
  group_by(compound) %>% #431 unique compounds
  summarize(state_cvr = n_distinct(state_name[!is.na(median_pest)])) |>
  mutate(pct_state_cvr = round((state_cvr/48)*100, 2))

#join median_pest and state-cover
dat_med_st_cvr<- left_join(median_pest, state_cover, by=c("compound"))

######################################################################
#Spatial coverage of chemicals applied (coverage that shows percent counties)
summ <- median_pest %>%
    group_by(state_name, compound) %>%
    summarize(n_counties_pest = ifelse(any(!is.na(median_pest)),
                                       n_distinct(county_fips_code[!is.na(median_pest)]), 0))

#number of counties
n_cnty<- median_pest %>%
  group_by(state_name) %>%
  summarize(num_counties = n_distinct(county_fips_code))

#join n_cnty with spatial coverage to get percent spatial coverage
summ_spat_pct<- left_join(summ,n_cnty, by="state_name") |>
  mutate(pct_spat_cover = round((n_counties_pest/num_counties)*100, 2))
  

#########################################################################

# Final data set that has median pesticide applied, pct spatial and temporal coverage for dat1 (2006-2019)

chem_appl_200619<- left_join(dat_med_st_cvr, summ_spat_pct, 
                      by=c("state_name","compound"))

chem_appl_200619<- chem_appl_200619 |>
  rename(state=state_name, pct_state_cvr = pct_state_cvr,
         pct_cnty_cover=pct_spat_cover) |>
  mutate(years = gsub("(^c\\(|\\)$)", "", years))

# save same file in rda format as backup
save(chem_appl_200619, file = "chem_appl_spat_temp_cvr2006_19.RDA")
```


#Step 3b chemical selection function
```{r}
compound_filter <- function(data, pct_yrs_thresh, pct_state_cvr_thresh, pct_cnty_cvr_thresh) {
  
  # Filter the data based on pct_yrs threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_yrs) >= pct_yrs_thresh)
  
  # Filter the data based on pct_state_cvr threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_state_cvr) >= pct_state_cvr_thresh)
  
  # Filter the data based on pct_cnty_cvr threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_cnty_cover) >= pct_cnty_cvr_thresh)
  
  # Return unique list of compounds that match the criteria as a dataframe
  unique(data$compound) %>% as.data.frame()
}
```

# Different percentages generate different lists
```{r}
#Use file: load("chem_appl_spat_temp_cvr2006_19.RDA")
compound_filter<- compound_filter(chem_appl_200619, 80, 80, 80)
compound_filter<- compound_filter |> rename(compound = ".")
#write_csv(compound_filter, "chemical_list90_90_90.csv")
```

####################################################################################
##################################################################################

# Step 4.1 Creating a final subset of dataset with selected chemicals for 1992-2005
```{r}
# read in the dataset
load("chem_appl_spat_temp_cvr92_2005.RDA")

# Filter chem_appl based on the values in compound_filter
chem_appl_filter9205 <- chem_appl_9205 %>%
  filter(compound %in% compound_filter$compound)

# Summarize pct_yrs by compound
summary <- chem_appl_filter9205 %>%
  group_by(compound) %>%
  summarize(
    mean_pct_yrs = mean(pct_cnty_cover),
    median_pct_yrs = median(pct_cnty_cover),
    max_pct_yrs = max(pct_cnty_cover),
    min_pct_yrs = min(pct_cnty_cover),
    n = n()
  )

save(chem_appl_filter9205, file="chem_appl_filter9205.RDA")
```

# Step 5 Grouping the pesticides by class and creating final dataset
```{r}

# create a named vector of compound classes
compound_classes <-list(Herbicide = c("2,4-D", "ATRAZINE", "ACETOCHLOR","CLETHODIM", "DICAMBA","GLYPHOSATE", 
                                      "IMAZETHAPYR", "PARAQUAT", "PENDIMETHALIN","CHLORIMURON","METOLACHLOR",
                                      "METRIBUZIN","METSULFURON", "NICOSULFURON", "PICLORAM", "SETHOXYDIM", 
                                      "SIMAZINE","THIFENSULFURON","TRIBENURON METHYL", "TRICLOPYR", "TRIFLURALIN"),
                        
                        Insecticide = c("CARBARYL", "CHLORPYRIFOS", "CYHALOTHRIN-LAMBDA",
                                        "DIMETHOATE", "ESFENVALERATE","PERMETHRIN"),
                        
                        Fungicide = c("CHLOROTHALONIL","MANCOZEB","PROPICONAZOLE"))

# Create a new variable "class" in chem_appl_filtered
chemi_class <- chem_appl_filtered %>% 
  mutate(class = case_when(
    compound %in% compound_classes[["Herbicide"]] ~ "Herbicide",
    compound %in% compound_classes[["Insecticide"]] ~ "Insecticide",
    compound %in% compound_classes[["Fungicide"]] ~ "Fungicide",
    TRUE ~ NA_character_
  ))

# Convert "class" variable to factor
chemi_class$class <- as.factor(chemi_class$class)

# Print the first few rows of the resulting dataset
head(chemi_class)

write_csv(chemi_class, "final_chemv5.csv")

# save same file in rda format as backup
save(chemi_class, file = "final_chem_v5_backup.RDA")
```
