==> 02_birth_defects_cleaning.Rmd <==
---
title: "02_birth_defect_cleaning"
output: pdf_document
date: "2023-03-24"
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, stringr, zipcodeR, lubridate, tmap, sf, tigris)
```


```{r}
benign_neop_mass<- list(155000,159800,162800,162801,162802,162803,162804,171800,186000,186001,186002,186003,
186004,189000,189001,189002,189003,189004,190500,190501,190502,190503,190504,191000,
194000,202300,208000,214000,214001,214002,214003,214004,214100,214200,214300,214400,
214401,214402,214403,214404,214800,214810,214900, 216910,216920, 238000,238010,238020,238030,
238040,238080,239200,239201,239202,239203,239204)
#count 51
lymphatic_malf<- list(228000, 228010,228020,228030,228031,228032,228033,228034,228040,228090,
                      228100,228101,228102,228103,228104,453000, 457800)
#count 2,306
eye<- list(237700, 330100,331890,335000,345600,351000,352600,362600,
           362601,362602,362603,362604,362700,363200,363201,363202,
           363203,363204,368000,378000,379500,740000,740010,740020,
           740030,740080,740100,740200,740210,740290,741000,741010,
           741020,741030,741040,741050,741060,741070,741080,741085,
           741086,741087,741090,741900,741910,741920,741930,741940,
           741980,741985,741990,742000,742080,742085,742086,742090,
           742100,742200,742210,742220,742230,742235,742240,742250,
           742260,742270,742280,742290,742300,742310,742320,742380,
           742390,742400,742410,742411,742412,742413,742414,742420,
           742421,742422,742423,742424,742480,742481,742482,742483,
           742484,742485,742486,742500,742510,742520,742530,742540,
           742580,742800,742810,742880,742900,742910,742990,743000,
           743001,743002,743003,743004,743100,743101,743102,743103,
           743104,743200,743201,743202,743203,743204,743210,743211,
           743212,743213,743214,743220,743221,743222,743223,743224,
           743300,743301,743302,743303,743304,743310,743311,743312,
           743313,743314,743320,743321,743322,743323,743324,743325,
           743326,743330,743331,743332,743333,743334,743340,743341,
           743342,743343,743344,743380,743381,743382,743383,743384,
           743390,743391,743392,743393,743394,743400,743401,743402,
           743403,743404,743410,743411,743412,743413,743414,743420,
           743421,743422,743423,743424,743430,743431,743432,743433,
           743434,743440,743441,743442,743443,743444,743480,743481,
           743482,743483,743484,743490,743491,743492,743493,743494,
           743500,743501,743502,743503,743504,743510,743511,743512,
           743513,743514,743520,743521,743522,743523,743524,743530,
           743531,743532,743533,743534,743535,743580,743581,743582,
           743583,743584,743590,743591,743592,743593,743594,743600,
           743601,743602,743603,743604,743610,743611,743612,743613,
           743614,743620,743621,743622,743623,743624,743630,743631,
           743632,743633,743634,743635,743636,743640,743641,743642,
           743643,743644,743650,743660,743661,743662,743663,743664,
           743670,743671,743672,743673,743674,743800,743801,743802,
           743803,743804,743810,743811,743812,743813,743814,743900,
           743901,743902,743903,743904,744020,744021,744022,744023,
           744024,744030,744031,744032,744033,744034,744090,744091,
           744092,744093,744094,756100)
#count:99
endo<- list(243990,251200,252100,253280,253820,255200,257800,759100,759101,759102,
            759103,759104,759110,759111,759112,759113,759114,759120,759121,759122,
            759123,759124,759130,759131,759132,759133,759134,759180,759181,759182,
            759183,759184,759190,759191,759192,759193,759194,759200,759210,759230,
            759231,759232,759233,759234,759280,759290,759300)
#count 1,595
genetic<- list(270100,270200,270300,270600,270700,271000,272700,275330,277000,277010,277500,
               277510,277620,277630,279110,758000,758010,758020,758030,
               758040,758090,758100,758110,758120,758130,758190,758200,
               758210,758220,758230,758290,758295,758300,758310,758320,
               758330,758340,758350,758360,758370,758380,758390,758400,
               758500,758510,758520,758530,758540,758580,758585,758586,
               758590,758600,758610,758690,758700,758710,758790,758800,
               758810,758820,758830,758840,758850,758860,758880,
               758890,758900,758910,758920,758930,758990,759340,759500,
               759600,759610,759620,759630,759680,759690,759700)
#count 1,998
hepatic<- list(277400,453000,750280,750300,750310,750320,750325,750330,750340,750350,
               750380,750400,750410,750420,750430,750480,750500,750510,750580,750600,
               750700,750710,750720,750730,750740,750750,750780,750800,750900,750910,
               750920,750990,751000,751010,751100,751110,751120,751190,751195,751200,
               751210,751220,751230,751240,751300,751310,751320,751330,751340,751400,
               751410,751420,751490,751495,751500,751510,751520,751530,751540,751550,
               751555,751560,751580,751590,751600,751610,751620,751630,751640,751650,
               751660,751670,751680,751700,751710,751720,751730,751740,751780,751790,
               751800,751810,751820,751880,751900,777100,777600,778000)
#count: 2
immune<- list(279200,759240)
#count: 6
hematolo<- list(282000,282100,282200,282600,284000,286000,286400) 
#cardiac count: 5,969
cardiac<- list(425300, 426705,427900,745000,745010,745100,745110,745120,745130,745140,745150,
               745180,745190,745200,745210,745300,745400,745410,745420,745480,745485,745486,
               745487,745490,745500,745510,745520,745570,745580,745590,745600,745610,45620,
               745630,745680,745690,745700,745800,745900,746000,746010,746020,746080,746090,
               746100,746105,746106,746200,746300,746400,746470,746480,746490,746500,746505,
               746600,746700,746800,746810,746820,746830,746840,746850,746860,746870,746880,
               746881,746882,746883,746885,746886,746887,746900,746910,746920,746930,746990,
               746995,747000,747100,747110,747120,747190,747200,747210,747215,747216,747217,
               747220,747230,747240,747250,747260,747270,747280,747285,747290,747300,747301,
               747302,747303,747304,747310,747320,747321,747322,747323,747324,747325,747330,
               747340,747380,747381,747382,747383,747384,747390,747391,747392,747393,747394,
               747400,747410,747420,747430,747440,747450,747480,747490,747500,747600,747601,
               747602,747603,747604,747610,747611,747612,747613,747614,747620,747630,747640,
               747641,747642,747643,747644,747650,747680,747690,747800,747810,747880,747900)
#cranial count: 6,934
cranial<-list(520600,524000,524080,744000,744001,744002,744003,744004,744010,744011,744012,744013,
              744014,744100,744110,744120,744200,744201,744202,744203,744204,744210,744211,744212,
              744213,744214,744220,744230,744231,744232,744233,744234,744240,744241,744242,744243,
              744244,744245,744246,744250,744251,744252,744253,744254,744280,744281,744282,744283,
              744284,744300,744301,744302,744303,744304,744400,744401,744402,744403,744404,744410,
              744480,744481,744482,744483,744484,744500,744800,744810,744820,744830,744880,744881,
              744882,744883,744884,744900,744910,748000,748001,748002,748003,748004,748100,748101,
              748102,748103,748104,748110,748120,748121,748122,748123,748124,748130,748140,748180,
              748181,748182,748183,748184,748185,748190,750000,750100,750110,750120,750130,750140,
              750180,750190,750200,750210,750230,750240,750250,750260,750270,754000,754010,754020,
              754030,754040,754050,754051,754052,754053,754054,754055,754060,754070,754080,754090,
              754100,754101,754102,754103,754104,756000,756001,756002,756003,756004,756005,756006,
              756010,756011,756012,756013,756014,756020,756021,756022,756023,756024,756030,756040,
              756045,756046,756050,756055,756056,756057,756060,756065,756080,756085,756090,759800,
              759220,749000,749010,749020,749030,749040,749050,749060,749070,749080,749090,749100,
              749110,749120,749190,749200,749210,749220,749290)
#abdominal count: 365
abdominal<- list(550000,550001,550002,550003,550004,550100,550101,550102,550103,550104,550900,550901,
                 550902,550903,550904,553100,553200,756700,756710,756720,756790,756795)

#lung count: 1,327
lung<- list(748205,748206,748209,748300,748310,748330,748340,748341,748342,748343,748344,748350,
            748351,748352,748353,748354,748360,748380,748385,748390,748400,748401,748402,748403,
            748410,748411,748412,748413,748414,748420,748421,748422,748423,748424,748480,748481,
            748482,748483,748484,748500,748501,748502,748503,748504,748510,748511,748512,748513,
            748514,748520,748521,748522,748523,748524,748580,748581,748582,748583,748584,748590,
            748591,748592,748593,748594,748600,748601,748602,748603,748604,748610,748611,748612,
            748613,748614,748620,748621,748622,748623,748624,748625,748690,748691,748692,748693,
            748694,748800,748801,748802,748803,748804,748810,748880,748881,748882,748883,748884,
            748900,756600,756610,756615,756616,756617,756620,756621,756622,756623,756624,756680,
            756681,756682,756683,756684,756690,756691,756692,756693,756694)

#renal/genitourinary: 4,106
renal_genuri<-list(752000,752001,752002,752003,752004,752010,752011,752012,752013,752014,752020,752021,
                   752022,752023,752024,752080,752081,752082,752083,752084,752085,752090,752091,752092,
                   752093,752094,752100,752101,752102,752103,752104,752110,752120,752121,752122,752123,
                   752124,752190,752191,752192,752193,752194,752200,752300,752310,752320,752380,752390,
                   752400,752410,752420,752430,752440,752450,752460,752470,752480,752490,752500,752501,
                   752502,752514,752520,752530,752531,752532,752533,752534,752600,752605,752606,752607,
                   752610,752620,752621,752625,752626,752627,752700,752710,752720,752730,752790,752800,
                   752801,752802,752803,752804,752810,752811,752812,752813,752814,752820,752830,752831,
                   752832,752833,752834,752840,752850,752860,752865,752870,752880,752900,753000,753009,
                   753010,753011,753012,753013,753100,753101,753102,753103,753110,753111,753112,753113,
                   753114,753120,753121,753122,753123,753124,753130,753131,753132,753133,753134,753140,
                   753141,753142,753143,753144,753150,753151,753152,753153,753154,753160,753161,753162,
                   753163,753164,753180,753181,753182,753183,753184,753200,753201,753202,753203,753204,
                   753210,753211,753212,753213,753214,753220,753221,753222,753223,753224,753290,753291,
                   753292,753293,753294,753300,753301,753302,753303,753304,753310,753311,753312,753313,
                   753314,753320,753321,753322,753323,753324,753330,753331,753332,753333,753334,753340,
                   753341,753342,753343,753344,753350,753351,753352,753353,753354,753380,753381,753382,
                   753383,753384,753400,753401,753402,753403,753404,753410,753411,753412,753413,753414,
                   753420,753421,753422,753423,753424,753480,753481,753482,753483,753484,753485,753500,
                   753600,753610,753620,753630,753690,753700,753710,753790,753800,753810,753820,753830,
                   753840,753850,753860,753870,753880,753900,753901,753902,753903,753904,753910,753911,
                   753912,753913,753914,753920,753930,753990,778600)

#musculoskeletal count: 4,350
musc_ske<- list(755021,755022,755023,755024,755030,755031,755032,755033,755034,755090,755091,
                755092,755093,755094,755095,755096,755100,755101,755102,755103,755104,755110,
                755111,755112,755113,755114,755120,755121,755122,755123,755124,755130,755131,
                755132,755133,755134,755190,755193,755196,754200,754210,754220,754300,754301,
                754302,754303,754304,754310,754311,754312,754313,754314,754400,754401,754402,
                754403,754404,754410,754411,754412,754413,754414,754420,754421,754422,754423,
                754424,754430,754431,754432,754433,754434,754440,754441,754442,754443,754444,
                754490,754491,754492,754493,754494,754500,754501,754502,754503,754504,754510,
                754511,754512,754513,754514,754520,754530,754531,754532,754533,754534,754590,
                754591,754592,754593,754594,754600,754601,754602,754603,754604,754610,754611,
                754612,754613,754614,754615,754680,754681,754682,754683,754684,754690,754691,
                754692,754693,754694,754700,754701,754702,754703,754704,754720,754721,754722,
                754723,754724,754730,754731,754732,754733,754734,754735,754780,754781,754782,
                754783,754784,754800,754810,754820,754825,754830,754831,754832,754833,754834,
                754840,754841,754842,754843,754844,754850,754851,754852,754853,754854,754880,
                754881,754882,754883,754884,755200,755201,755202,755203,755204,755210,755211,
                755212,755213,755214,755220,755221,755222,755223,755224,755230,755231,755232,
                755233,755234,755240,755241,755242,755243,755244,755250,755251,755252,755253,
                755254,755260,755261,755262,755263,755264,755265,755270,755271,755272,755273,
                755274,755280,755281,755282,755283,755284,755285,755290,755291,755292,755293,
                755294,755300,755301,755302,755303,755304,755310,755311,755312,755313,755314,
                755320,755321,755322,755323,755324,755330,755331,755332,755333,755334,755340,
                755341,755342,755343,755344,755350,755351,755352,755353,755354,755360,755361,
                755362,755363,755364,755365,755366,755380,755381,755382,755383,755384,755385,
                755390,755391,755392,755393,755394,755400,755401,755402,755403,755404,755410,
                755411,755412,755413,755414,755420,755421,755422,755423,755424,755430,755431,
                755432,755433,755434,755440,755441,755442,755443,755444,755480,755481,755482,
                755483,755484,755490,755491,755492,755493,755494,755500,755501,755502,755503,
                755504,755510,755511,755512,755513,755514,755520,755521,755522,755523,755524,
                755525,755526,755530,755531,755532,755533,755534,755535,755536,755540,755541,
                755542,755543,755544,755550,755551,755552,755553,755554,755555,755556,755560,
                755561,755562,755563,755564,755580,755581,755582,755583,755584,755585,755590,
                755591,755592,755593,755594,755600,755601,755602,755603,755604,755605,755606,
                755610,755611,755612,755613,755614,755616,755620,755621,755622,755623,755624,
                755630,755631,755632,755633,755634,755640,755641,755642,755643,755644,755645,
                755646,755647,755650,755651,755652,755653,755654,755660,755661,755662,755663,
                755664,755665,755670,755680,755681,755682,755683,755684,755685,755690,755691,
                755692,755693,755694,755800,755810,755880,755900,756110,756120,756130,756140,
                756145,756146,756150,756155,756156,756160,756165,756166,756170,756175,756179,
                756180,756185,756190,756200,756300,756301,756302,756303,756304,756310,756311,
                756312,756313,756314,756320,756321,756322,756323,756324,756330,756331,756332,
                756333,756334,756340,756341,756342,756343,756344,756350,756360,756380,756390,
                756400,756410,756420,756430,756445,756446,756447,756450,756460,756470,756480,
                756490,756500,756505,756506,756510,756520,756525,756530,756540,756550,756560,
                756570,756575,756580,756590,756800,756801,756802,756803,756804,756810,756811,
                756812,756813,756814,756820,756821,756822,756823,756824,756830,756840,756850,
                756860,756861,756862,756863,756864,756880,756900,756910,756920,756930,756940,
                756990,759820,759840,759860,759870)

#skin and soft tissue: 1,602
skin_soft_tiss<- list(685100,757000,757001,757002,757003,757004,757100,757110,757115,757120,
                      757190,757195,757196,757197,757200,757300,757310,757320,757330,757340,
                      757345,757346,757350,757360,757370,757380,757390,757395,757400,757410,
                      757420,757430,757450,757480,757500,757501,757502,757503,757504,757510,
                      757511,757512,757513,757514,757515,757516,757520,757530,757540,757541,
                      757542,757543,757544,757580,757585,757600,757601,757602,757603,757604,
                      757610,757611,757612,757613,757614,757620,757621,757622,757623,757624,
                      757630,757631,757632,757633,757634,757640,757641,757642,757643,757644,
                      757650,757680,757681,757682,757683,757684,757800,757900,757910,757920,
                      757990)

#other count: 470
other<- list(658800,759000,759005,759010,759020,759030,759040,759041,759042,759043,759044,
             759050,759051,759052,759053,759054,759080,759090,759310,759320,759330,759390,
             759400,759410,759420,759430,759440,759480,759490,759890,759900,759910,759990,
             999999)
#unlikely: 30
unlike<- list(608200,608201,608202,608203,608204,760710,760750,767600,771000,771090,771100,
              771210,771220,771230,771280,774480,774490,90000)

```

===============================================================================
===============================================================================

```{r}
data_bd <- read_csv("birth_def_v1.csv") |>
  mutate(city = str_trim(city)) |>
  mutate(city = str_to_title(city))

# Create a list of objects to filter by
filter_list <- list(pulmonary = lung, cardiac = cardiac, hepatic = hepatic,
                    skin_soft_tiss=skin_soft_tiss, musc_ske=musc_ske,
                    renal_genuri=renal_genuri, abdominal=abdominal,
                    cranial=cranial, hematolo=hematolo,immune=immune,
                    genetic=genetic, endocrine=endo, ophthal=eye,
                    lymphatic=lymphatic_malf, benign_neop_mass=benign_neop_mass)

# Use map() function to apply filter_all() to each object in filter_list
dt_list <- map(filter_list, ~ data_bd |> filter_all(any_vars(. %in% .x)))

# Rename the list elements to include "dat_"
names(dt_list) <- paste0("dat_", names(dt_list))

# Save each data frame to its own object in the global environment
list2env(dt_list, envir = .GlobalEnv)
```

```{r}
# Find all objects in the global environment that start with "dat_"
dt_names <- ls(pattern = "^dat_")

# Combine all data frames into a single data frame
```


```{r}
dt_combined <- bind_rows(mget(dt_names), .id = "df_name")

# Extract the dataframe name followed by "_" and add a new variable "class"
dt_combined <- dt_combined %>% 
  mutate(class = gsub("^dat_(.*)", "\\1", df_name))

write_csv(dt_combined, "birth_defects_v2.csv")
```

#Reference zip-code file
```{r}
zip_dat<- zipcodeR::search_state("NE") |> select(c(1,3,6)) |>
  rename(zip9= "zipcode") |>
  mutate(zip9=as.factor(zip9))

birth_defects<- birth_defects |> left_join(zip_dat, by="zip9")
write_csv(birth_defects, "birth_defects_v2.csv")

dt_combined<- read_csv("birth_defects_v2.csv") |>
  filter(is.na(dth_cert)) |>
  group_by(bth_cert, case, bwt_gm, sexc, fac, city) |>
  filter(n() == 1)

write_csv(dt_combined, "birth_defects_v3.csv")
```
#==============================================================================#
#==============================================================================#
#==============================================================================#
#adding county infomation to the birth defects dataset
```{r}
dt_bd<- read_csv("birth_defects_v3.csv")
dt_bd$zip9 <- as.factor(dt_bd$zip9)

zip_dat<- zipcodeR::search_state("NE") |> select(c(1,3,6)) |>
  rename(zip9= "zipcode") |>
  mutate(zip9=as.factor(zip9))

#birth defects 
dt_bd_v4 <- dt_bd |> left_join(zip_dat, by="zip9")

#use nebraska zipcodes to drop any bd observations not from nebraska
filtered_dt_bd_v4 <- dt_bd_v4 %>%
  semi_join(lb, by = "county")

#final birth defects dataset for any individual level analysis
write_csv(filtered_dt_bd_v4, "birth_defects_v4.csv")

#use the same dt_db_v4 for generating the county level counts of birth defects
bd_county<- filtered_dt_bd_v4 |>
  select(c(63,65)) |>
  group_by(county, class) |>
  summarise(observations = n()) |>
  pivot_wider(names_from = class, values_from = observations) |>
  mutate(total_bd = sum(c_across(where(is.numeric)), na.rm = TRUE))

#final birth defects dataset for any county level analysis
write_csv(bd_county, "birth_defects_county.csv") 

```


#adding county information to live births data
```{r}
live_birth <- read_csv("live_birth_ne.csv")|>
  clean_names() |>
  rename(zip9= "mother_resident_zip") |>
  mutate(zip9=as.factor(zip9)) |>
  filter(mother_resident_state == "Nebraska")


zip_dat<- zipcodeR::search_state("NE") |> select(c(1,3,6)) |>
  rename(zip9= "zipcode") |>
  mutate(zip9=as.factor(zip9))

live_birth<- live_birth |> left_join(zip_dat, by="zip9") |>
  drop_na()
write_csv(live_birth, "live_birth_v1.csv")

#### aggregating live births
lb<- read_csv("live_birth_v1.csv") |>
  select(c(6)) |>
  count(county)

write_csv(lb, "live_birth_v2.csv")
```

##############################################################################
#combine live birth and birth defect counts per county
```{r}
bd<- read_csv("birth_defects_county.csv")
lb<- read_csv("live_birth_v2.csv")

bd_county_dat<- lb |>
  left_join(bd, by= "county") |>
  rename(live_birth= "n")


write_csv(bd_county_dat, "bd_cnty_v7.csv")
```


# Birth defects prevalence rate
```{r}

bd_prev_dat <- read_csv("bd_cnty_v7.csv") |>
  mutate(across(where(is.integer), as.numeric)) |>
  mutate(across(3:17, ~ . / live_birth)*100000)

write_csv(bd_prev_dat, "bd_prev_v8.csv")
```


# joining birth defects_live birth data with NE counties shapefile 
```{r}
bd_dt <- read_csv("bd_prev_v8.csv")

ne_counties <- bd_dt |>
  counties(state = "ne", cb = FALSE, resolution = "500k", year = "2010", class = "sf") |>
  clean_names()

#join birth defects_live birth data with ne county shapefile
bd_live_ne_map<- bd_dt %>%
 left_join(ne_counties, by = c("county" = "namelsad10"))

write_csv(bd_live_ne_map, "bd_cnty_v9.csv")

```


# barplots by birth defect types
```{r}
dat_bd_plt<- read_csv("bd_cnty_v9.csv")  |>
  select(-c(2,18:36))|>
  pivot_longer(cols = -c(county),
               names_to = "bd_type", values_to = "bd_rate") |>
  mutate(bd_type_lable= substr(bd_type, 1, 5))

ggplot(dat_bd_plt, aes(x = bd_rate, y = county)) +
  geom_bar(stat = "identity") +
  labs(title = "", x = "", y = "") +
  theme_minimal()+
  facet_grid(.~bd_type_lable, scales = "free") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

```

# Map birth defects
```{r}
bd_dt <- read_csv("bd_prev_v8.csv")

ne_counties <- bd_dt %>%
  counties(state = "ne", cb = FALSE, resolution = "500k", year = "2010", class = "sf") %>%
  clean_names()

bd_live_ne_map <- bd_dt %>%
  left_join(ne_counties, by = c("county" = "namelsad10")) %>%
  st_as_sf()  # Convert to sf class

w <- tm_shape(bd_live_ne_map) +
  tm_polygons(col = "cardiac",
              breaks = quantile(bd_live_ne_map$cardiac, probs = c(0, 0.25, 0.5, 0.75, 1), na.rm = TRUE),
              legend.is.portrait = TRUE,
              palette = "viridis",
              title = "Cardiac defects per 100k children") +
  tm_borders() +
   tm_text(text = "county", size= 0.6) +
  tm_borders(col = "blue", lwd = 2, alpha = 0.5)+
    tm_compass(position = c("right", "top"))

tmap_save(w,"ne_cardiacdefect_viridis_palette.tiff",
          width = 12, height = 15, dpi = 150)

```

#Summary table for manuscript
```{r}
# 1. Categorize gestation age; read birth_def_v1.csv
data_bd <- data_bd %>%
  mutate(gest_category = if_else(gest < 37, "<37", ">=37"))

# Get the frequency of each category; <37= 5111 & >=37 = 19,854
gest_freq <- data_bd %>%
  group_by(gest_category) %>%
  summarise(frequency = n())

# 2. Categorize birth weight
data_bd <- data_bd %>%
  mutate(bw_category = if_else(bwt_gm <2500, "<2500", ">=2500"))

# Get the frequency of each birth weight category; <2500 =5062  & >=2500 = 19901 and NA = 2
bw_freq <- data_bd %>%
  group_by(bw_category) %>%
  summarise(frequency = n())

# 3. Categorize birth year
data_bd <- data_bd %>%
  mutate(byr_category = if_else(bth_yr <2010, "<2010", ">=2010"))

# Get the frequency of each birth weight category; <2010= 16,891  & >=2010= 8074  
byr_freq <- data_bd %>%
  group_by(byr_category) %>%
  summarise(frequency = n())

# 4. Categorize maternal age
data_bd <- data_bd %>%
  mutate(mage_category = if_else(mage <35, "<35", ">=35"))

# Get the frequency of each birth weight category; <35= 21474 & >=35= 3486, NA = 5
mage_freq <- data_bd %>%
  group_by(mage_category) %>%
  summarise(frequency = n())

# 5. Get the frequency of each value in the hispm variable; 1= 1460, 2= 42, 5 = 1318, 6= 22106 (Refer natality code in dissertation data onedrive folder)
hispm_freq <- data_bd %>%
  count(hispm) %>%
  filter(hispm >= 1 & hispm <= 6)

# 6. Get the frequency of each value in the mrace variable; 0 = 1877; 1= 20707; 2=1578; 3= 247; 4= 32; 5= 18; 6= 5; 7= 28; 8= 450; 9= 23;
#9 categories: American Indian or Alaska Native; Black; Chinese; Filipino; Guamanian; Hawaiian (includes part-Hawaiian); Japanese; Other Asian; White.
mrace_freq <- data_bd %>%
  count(mrace) %>%
  filter(mrace >= 0 & mrace <= 9)

#7. Get the frequency of each value in the sexc variable/infant gender; F= 10298; M= 14666
sexc_freq <- data_bd %>%
  count(sexc)

#8. Get the frequency of each value in the diabetes variable; Y= 1151; N= 23763; Unknown= 51
diabetes_freq <- data_bd %>%
  count(diabetes)

#9. Get the frequency of each value in the tobacco variable; Y= 4328; N= 20615; Unknown= 22
tobacco_freq <- data_bd %>%
  count(tobacco)

```
