==> 02.1_nebraska_pest_analysis.Rmd <==
---
title: "02.1_nebraska_pest_analysis"
author: "Taiba"
date: "2023-06-06"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor,reshape2)
```

# 1 Loading original conus data 1992-2019
```{r}

load("~/Desktop/Jabbeen/jt_dissert/jt_dissert/pest_conus_v2.RDA")
dat_conus$year<- as.numeric(as.character(dat_conus$year))
```

# Step 1 Filtering pesticides for NE from 1992- 2014 has 205956 obs and 6 variables
```{r}
dat_ne <- dat_conus |>
  filter(state_fips_code == "31" & year %in% c(1992, 1993, 1994, 1995, 1996, 1997, 1998, 1999, 2000,
                                               2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009,
                                               2010, 2011, 2012, 2013, 2014))

  save(dat_ne, file = "pest_ne_v1.RDA")
```

# Step 2 creating variables median, temporal and spatial variables using original dataset 03/30/23
```{r}
load("pest_ne_v1.RDA")
#median quantity and temporal coverage of chemicals applied
median_pest <- dat_ne %>%
  group_by(county_fips_code, compound) %>%
  summarize(median_pest = median(pest_applied, na.rm = TRUE),
            n_yrs_repeat = sum(!is.na(pest_applied)),
            years = list(unique(year[!is.na(pest_applied)])))|>
  mutate(pct_yrs= round((n_yrs_repeat/23)*100, 2))


#spatial coverage using counties (percent counties with any compound applied per year)
county_cover <- median_pest %>%
  group_by(compound) %>%
  summarize(county_cvr = n_distinct(county_fips_code[!is.na(median_pest)])) |>
  mutate(pct_county_cvr = round((county_cvr/93)*100, 2))

#join median_pest and county-cover
dat_med_cnty_cvr<- left_join(median_pest, county_cover, by=c("compound"))

#########################################################################

# Final data set that has median pesticide applied, pct spatial and temporal coverage has 18730 obs and 8 variables

chem_appl<- dat_med_cnty_cvr |>
  mutate(years = gsub("(^c\\(|\\)$)", "", years))

# save same file in rda format as backup
save(chem_appl, file = "ne_chem_appl_spat_temp_cvr_v2.RDA")
```

#Step 3 chemical selection function
```{r}
compound_filter <- function(data, pct_yrs_thresh, pct_cnty_cvr_thresh) {
  
  # Filter the data based on pct_yrs threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_yrs) >= pct_yrs_thresh)
  
  # Filter the data based on pct_cnty_cvr threshold
  data <- data %>% group_by(compound) %>% filter(mean(pct_county_cvr) >= pct_cnty_cvr_thresh)
  
  # Return unique list of compounds that match the criteria as a dataframe
  unique(data$compound) %>% as.data.frame()
}
```

# Different percentages generate different lists
```{r}
#Use file: load("ne_chem_appl_spat_temp_cvr_v2.RDA")
compound_filter<- compound_filter(chem_appl,90,100)
compound_filter<- compound_filter |> rename(compound = ".")
write_csv(compound_filter, "ne_chemical_list90_100.csv")

```


# Step 4 Creating a final subset of dataset with selected chemicals 
```{r}
# read in the dataset
load("ne_chem_appl_spat_temp_cvr_v2.RDA")

# Filter chem_appl based on the values in compound_filter
chem_appl_filtered <- chem_appl %>%
  filter(compound %in% compound_filter$compound)

# Summarize pct_yrs by compound
summary <- chem_appl_filtered %>%
  group_by(compound) %>%
  summarize(
    mean_pct_yrs = mean(pct_county_cvr),
    median_pct_yrs = median(pct_county_cvr),
    max_pct_yrs = max(pct_county_cvr),
    min_pct_yrs = min(pct_county_cvr),
    n = n()
  )

save(chem_appl_filtered, file="ne_chem_appl_filtered_v3.RDA")
```

# Step 5 Grouping the pesticides by class and creating final dataset has 2976 obs and 9 variables
```{r}

# create a named vector of compound classes
compound_classes <-list(Herbicide = c("2,4-D", "ATRAZINE","ALACHLOR","ACETOCHLOR","BROMOXYNIL", "CLETHODIM","CLOPYRALID", "DICAMBA","DIMETHENAMID","FLUMETSULAM","GLYPHOSATE", 
                                      "IMAZETHAPYR","MCPA","METOLACHLOR","PARAQUAT", 
                                      "METRIBUZIN","METSULFURON", "NICOSULFURON", "PICLORAM","PENDIMETHALIN","QUIZALOFOP", "SETHOXYDIM", "THIFENSULFURON","TRIASULFURON", "TRIFLURALIN"),
                        
                        Insecticide = c( "BIFENTHRIN" ,"CHLORPYRIFOS", "DIMETHOATE", "ESFENVALERATE","PERMETHRIN","TEFLUTHRIN","TERBUFOS"))

# Create a new variable "class" in chem_appl_filtered
chemi_class <- chem_appl_filtered %>% 
  mutate(class = case_when(
    compound %in% compound_classes[["Herbicide"]] ~ "Herbicide",
    compound %in% compound_classes[["Insecticide"]] ~ "Insecticide",
    TRUE ~ NA_character_
  ))

# Convert "class" variable to factor
chemi_class$class <- as.factor(chemi_class$class)

# Print the first few rows of the resulting dataset
head(chemi_class)

write_csv(chemi_class, "ne_final_chemv4.csv")

# save same file in rda format as backup
save(chemi_class, file = "ne_final_chem_v4_backup.RDA")

```


# Step 6 boxplots for each chemical
```{r}
load("ne_final_chem_v4_backup.RDA")

ggplot(chemi_class, aes(y = compound, x = log(median_pest))) +
  geom_boxplot() +
  labs(title = "",
       x = "Median pesticide applied (1992 - 2014) - Log scale", 
       y = "") +
  facet_grid(factor(class, levels=c("Herbicide", "Insecticide"))~ ., scales = "free_y", 
             switch = "y", space = "free_y") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 0, hjust = 1, size = 10.5),
        axis.text.y = element_text(size = 10.5),
        strip.text = element_text(colour = "black", size = 10.5),
        strip.text.x = element_text(colour = "black", size = 10.5),
        strip.text.y = element_text(colour = "black", size = 7))


ggsave("ne_chemical_box_plt.tiff", 
       width = 6,height = 9,
       dpi=300)
```

#Step 7 correlation matrix
```{r}
# pivot from long to wide format
chemi_class_wide <- chemi_class[c(1:3)] %>%
  pivot_wider(names_from = compound, values_from = median_pest)|>
   mutate_at(vars(2:33), ~log10(.+0.0000001)) 

write_csv(chemi_class_wide, "ne_final_chem_widev5.csv")

# save same file in rda format as backup
save(chemi_class_wide, file = "ne_final_chem_wide_v5_backup.RDA")
```


```{r}
desired_order <- c("2,4-D", "Acetochlor", "Alachlor", "Atrazine", "Bromoxynil", 
                   "Clethodim", "Clopyralid", "Dicamba", "Dimethenamid",
                   "Flumetsulam", "Glyphosate", "Imazethapyr", "Mcpa",
                   "Metolachlor", 
                   "Metribuzin", "Metsulfuron", "Nicosulfuron", "Paraquat", "Pendimethalin",
                   "Picloram", "Quizalofop", "Sethoxydim",
                   "Thifensulfuron", 
                   "Triasulfuron", "Trifluralin",
                   "Bifenthrin", "Chlorpyrifos", "Dimethoate", "Esfenvalerate", "Permethrin", 
                   "Tefluthrin", "Terbufos")

chemi_class_wide <- read_csv("ne_final_chem_widev5.csv") |>
  rename_all(~str_to_title(.))
  
#correlation matrix
cormat<- round(x=cor(chemi_class_wide[c(2:33)], method = "spearman", 
                     use= "complete.obs"), digits=2) |>
  melt() |>
  mutate_at(vars(Var1, Var2), ~factor(., levels = rev(desired_order)))


ggplot(cormat, aes(x=Var2, y=rev(Var1), fill=value)) +
  geom_tile(color="white")+
  scale_fill_gradient2(low="red", high="blue", mid="white",
                       midpoint=0,
                       limit=c(-1,1), space= "Lab",
                       name="Spearman Correlation | Pesticides applied in NE Counties (1992-2014)")+
  geom_text(aes(label = round(value, 2)), size = 3, color = "black") + # Add text labels
  theme_minimal()+
  geom_hline(yintercept = 7.5, linetype = "dashed", color = "black", size=1.5)+
  geom_vline(xintercept = 7.5, linetype = "dashed", color = "black", size=1.5)+
  theme(axis.text.x = element_text(angle = 45, vjust = 1, size = 12, hjust = 1),
        axis.text.y = element_text(angle = 0, vjust = 1, size = 12, hjust = 1),
        axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        panel.grid.major = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank(),
        axis.ticks = element_blank(),
        legend.position = "bottom", legend.box = "horizontal")+
  coord_fixed()

ggsave("ne_chemical_correlations.tiff",
       width=12, height= 10, dpi=300)
```

#################################################################################
08/06/23 to check top chemicals for Dr. Rogan

 # 
 load (""ne_final_chem_v4_backup.RDA")
 
 #
 Sort the dataset based on the highest median_pest per compound
sorted_chemi_class <- chemi_class %>%
 arrange(compound, desc(median_pest))

# View the sorted dataset
print(sorted_chemi_class)

##############################################################################
11/08/23 creating a table on pesticide characteristics for manuscript

library(tidyverse)

 # 
 load ("ne_final_chem_v4_backup.RDA")

# Assuming 'chemi_class' is dataframe
chemi_class_grouped <- chemi_class %>%
  group_by(compound) %>%
  summarize(
    median = mean(median_pest, na.rm = TRUE), # Calculate the median by taking the average of median
    class = first(class), # Get the class for each compound
    .groups = 'drop' # Drop grouping information
  ) %>%
  mutate(
    median = round(median, 2), # Round median_pest to two decimal places
    rank = rank(-median) # Create the rank based on the rounded median
  ) %>%
  select(rank, compound, class, median) %>%
  arrange(rank)

# Print the data to see the result
print(chemi_class_grouped, n = 32)

# save as csv format

write_csv(chemi_class_grouped, "ne_final_chem_char.csv")
